refactor: Move the input modal to its own component
