+++ src/aux-server/aux-web/shared/vue-components/ShowInputModal/ShowInputModal.css
@@ -1,0 +1,27 @@
+.input-dialog-content {
+    padding-bottom: 0;
+}
+
+.input-dialog-content .md-field {
+    margin-bottom: 20px;
+}
+
+.input-dialog .md-dialog-title {
+    margin-bottom: 0;
+}
+
+.input-dialog-color-tools {
+    padding-bottom: 16px;
+}
+
+.input-dialog {
+    transition-duration: 0.1s, 0.1s;
+}
+
+.color-picker-advanced {
+    margin: auto;
+}
+
+.color-picker-basic.vc-compact {
+    width: 245px;
+}
